DCS-233 Extracting out reminder functionality from notification service.
