+++ server/services/jobs/notificationJobs.js
@@ -1,11 +1,11 @@
 const logger = require('../../../log.js')
 
-module.exports = function createNotificationJobs(notificationService, signInService) {
+module.exports = function createNotificationJobs(reminderService, signInService) {
   async function roReminders() {
     logger.info('Running RO reminders')
     try {
       const systemToken = await signInService.getAnonymousClientCredentialsTokens()
-      return await notificationService.notifyRoReminders(systemToken.token)
+      return await reminderService.notifyRoReminders(systemToken.token)
     } catch (error) {
       logger.error('Error running RO reminders', error.stack)
     }
