+++ server/views/admin/delius/managedOffenders.pug
@@ -5,25 +5,26 @@ block content
   div.back-link-container.smallPaddingTop
     a#back.link-back(href="/admin/delius") Back
 
-  div.pure-g.pure-u-1.paddingTop
-    form(method='POST' action='/admin/delius/responsibleOfficer')
-      input(type="hidden" name="_csrf" value=csrfToken)
-      input(type="hidden" name="offenderNo" value=offenders[0].offenderNo)
-      div.pure-u-1-3
-        input.link(type="submit" value="View responsible officer for these offenders")
+  if offenders.length
+    div.pure-g.pure-u-1.paddingTop
+      form(method='POST' action='/admin/delius/responsibleOfficer')
+        input(type="hidden" name="_csrf" value=csrfToken)
+        input(type="hidden" name="offenderNo" value=offenders[0].offenderNo)
+        div.pure-u-1-3
+          input.link(type="submit" value="View responsible officer for these offenders")
 
-  div.pure-g.pure-u-1
-    table.largeMarginBottom
-      thead
-        tr
-          th Booking id
-          th Offender no
-          th First name
-          th Last name
-      tbody
-        each offender in offenders
+    div.pure-g.pure-u-1
+      table.largeMarginBottom
+        thead
           tr
-            td #{offender.bookingId}
-            td #{offender.offenderNo}
-            td #{offender.firstName}
-            td #{offender.lastName}
+            th Booking id
+            th Offender no
+            th First name
+            th Last name
+        tbody
+          each offender in offenders
+            tr
+              td #{offender.bookingId}
+              td #{offender.offenderNo}
+              td #{offender.firstName}
+              td #{offender.lastName}
