+++ test/services/jobs/notificationJobsTest.js
@@ -2,24 +2,24 @@ const createNotificationJobs = require('../../../server/services/jobs/notificati
 
 describe('notificationJobs', () => {
   let signInService
-  let notificationService
+  let reminderService
   let jobs
 
   const notifyRoRemindersResult = { result: 'result' }
 
   beforeEach(() => {
     signInService = { getAnonymousClientCredentialsTokens: sinon.stub().resolves({ token: 'test-token' }) }
-    notificationService = { notifyRoReminders: sinon.stub().resolves(notifyRoRemindersResult) }
+    reminderService = { notifyRoReminders: sinon.stub().resolves(notifyRoRemindersResult) }
 
-    jobs = createNotificationJobs(notificationService, signInService)
+    jobs = createNotificationJobs(reminderService, signInService)
   })
 
   it('should sign in and trigger notifications', async () => {
     const result = await jobs.roReminders()
 
     expect(signInService.getAnonymousClientCredentialsTokens).to.be.calledOnce()
-    expect(notificationService.notifyRoReminders).to.be.calledOnce()
-    expect(notificationService.notifyRoReminders).to.be.calledWith('test-token')
+    expect(reminderService.notifyRoReminders).to.be.calledOnce()
+    expect(reminderService.notifyRoReminders).to.be.calledWith('test-token')
     expect(result).to.eql(notifyRoRemindersResult)
   })
 })
