+++ test/services/notificationServiceTest.js
@@ -5,14 +5,11 @@ describe('notificationService', () => {
   let service
   let prisonerService
   let userAdminService
-  let deadlineService
   let configClient
   let notifyClient
   let audit
   let nomisClient
   let config
-
-  const transitionDate = '2019-01-01 12:00:00'
 
   beforeEach(() => {
     config = {
@@ -48,15 +45,6 @@ describe('notificationService', () => {
         email: 'ro@ro.email',
         organisation: 'NPS Dewsbury (Kirklees and Wakefield)',
       }),
-    }
-    deadlineService = {
-      getDueInDays: sinon.stub().resolves([{ booking_id: 1, transition_date: transitionDate }]),
-      getOverdue: sinon
-        .stub()
-        .resolves([
-          { booking_id: 2, transition_date: transitionDate },
-          { booking_id: 3, transition_date: transitionDate },
-        ]),
     }
     configClient = {
       getMailboxes: sinon
@@ -97,7 +85,6 @@ describe('notificationService', () => {
     service = createNotificationService(
       prisonerService,
       userAdminService,
-      deadlineService,
       configClient,
       notifyClient,
       audit,
@@ -499,7 +486,6 @@ describe('notificationService', () => {
         service = createNotificationService(
           prisonerService,
           userAdminService,
-          deadlineService,
           configClient,
           notifyClient,
           audit,
@@ -710,114 +696,6 @@ describe('notificationService', () => {
 
       const result = await service.sendNotifications(inputs)
       expect(result).to.eql([])
-    })
-  })
-
-  describe('notifyRoReminders', () => {
-    it('should get notifiable booking IDs from deadline service', async () => {
-      await service.notifyRoReminders('token')
-      expect(deadlineService.getOverdue).to.be.calledOnce()
-      expect(deadlineService.getDueInDays).to.be.calledTwice()
-      expect(deadlineService.getDueInDays).to.be.calledWith('RO', 0)
-      expect(deadlineService.getDueInDays).to.be.calledWith('RO', 2)
-    })
-
-    it('should return counts of notifiable cases', async () => {
-      const result = await service.notifyRoReminders('token')
-      expect(result).to.eql({ overdue: 2, due: 1, soon: 1 })
-    })
-
-    it('should do nothing further if empty notifiable cases', async () => {
-      deadlineService.getDueInDays = sinon.stub().resolves()
-      deadlineService.getOverdue = sinon.stub().resolves()
-      const result = await service.notifyRoReminders('token')
-
-      expect(prisonerService.getOrganisationContactDetails).to.have.callCount(0)
-      expect(prisonerService.getPrisonerPersonalDetails).to.have.callCount(0)
-      expect(result).to.eql({ overdue: 0, due: 0, soon: 0 })
-    })
-
-    it('should do nothing further if no notifiable cases', async () => {
-      deadlineService.getDueInDays = sinon.stub().resolves([])
-      deadlineService.getOverdue = sinon.stub().resolves([])
-      await service.notifyRoReminders('token')
-
-      expect(prisonerService.getOrganisationContactDetails).to.have.callCount(0)
-      expect(prisonerService.getPrisonerPersonalDetails).to.have.callCount(0)
-    })
-
-    it('should get submissionTarget and prisonerDetails for each notifiable case', async () => {
-      await service.notifyRoReminders('token')
-      expect(prisonerService.getOrganisationContactDetails).to.have.callCount(4)
-      expect(prisonerService.getOrganisationContactDetails).to.be.calledWith('RO', 1, 'token')
-      expect(prisonerService.getOrganisationContactDetails).to.be.calledWith('RO', 2, 'token')
-      expect(prisonerService.getOrganisationContactDetails).to.be.calledWith('RO', 3, 'token')
-
-      expect(prisonerService.getPrisonerPersonalDetails).to.have.callCount(4)
-      expect(prisonerService.getPrisonerPersonalDetails).to.be.calledWith(1, 'token')
-      expect(prisonerService.getPrisonerPersonalDetails).to.be.calledWith(2, 'token')
-      expect(prisonerService.getPrisonerPersonalDetails).to.be.calledWith(3, 'token')
-    })
-
-    it('should call notify client for each notification', async () => {
-      const expectedCommonData = {
-        offender_dob: '1/1/1',
-        offender_name: 'First Last',
-        offender_noms: 'AB1234A',
-        domain: 'http://localhost:3000',
-      }
-
-      const firstNotification = {
-        ...expectedCommonData,
-        booking_id: 2,
-        date: 'Tuesday 15th January',
-        prison: 'HMP Blah',
-        ro_name: undefined,
-        organisation: 'NPS Dewsbury (Kirklees and Wakefield)',
-      }
-      const secondNotification = {
-        ...expectedCommonData,
-        booking_id: 3,
-        date: 'Tuesday 15th January',
-        prison: 'HMP Blah',
-        ro_name: undefined,
-        organisation: 'NPS Dewsbury (Kirklees and Wakefield)',
-      }
-      const overdueTemplate = templates.RO_OVERDUE.templateId
-      const expectedEmail = 'ro@ro.email'
-
-      await service.notifyRoReminders('token')
-
-      expect(notifyClient.sendEmail).to.have.callCount(10)
-      expect(notifyClient.sendEmail).to.be.calledWith(overdueTemplate, expectedEmail, {
-        personalisation: firstNotification,
-      })
-      expect(notifyClient.sendEmail).to.be.calledWith(overdueTemplate, expectedEmail, {
-        personalisation: secondNotification,
-      })
-    })
-
-    it('should continue with subsequent reminder types even after a failure', async () => {
-      deadlineService.getOverdue.rejects(new Error('error message'))
-      deadlineService.getDueInDays = sinon.stub().resolves()
-      await service.notifyRoReminders('token')
-
-      expect(deadlineService.getOverdue).to.be.calledOnce()
-      expect(deadlineService.getDueInDays).to.be.calledTwice()
-      expect(deadlineService.getDueInDays).to.be.calledWith('RO', 0)
-      expect(deadlineService.getDueInDays).to.be.calledWith('RO', 2)
-    })
-
-    it('should continue trying all other notifications even after a failure', async () => {
-      prisonerService.getOrganisationContactDetails.rejects(new Error('error message'))
-      await service.notifyRoReminders('token')
-
-      expect(deadlineService.getOverdue).to.be.calledOnce()
-      expect(deadlineService.getDueInDays).to.be.calledTwice()
-      expect(deadlineService.getDueInDays).to.be.calledWith('RO', 0)
-      expect(deadlineService.getDueInDays).to.be.calledWith('RO', 2)
-      expect(prisonerService.getOrganisationContactDetails).to.have.callCount(4)
-      expect(prisonerService.getPrisonerPersonalDetails).to.have.callCount(4)
     })
   })
 })
